shopify-image-translator/requirements.txt
Flask
Pillow
google-cloud-vision
google-cloud-translate
requests
python-dotenv
